move vulkan test to separate file so CI skips it
